+++ rules/deprecated/typescript.js
@@ -1,0 +1,11 @@
+'use strict';
+/**
+ * The following @typescript-eslint rules have been deprecated.
+ *
+ */
+
+module.exports = {
+  rules: {
+    '@typescript-eslint/ban-ts-ignore': 'off',
+  },
+};