+++ CHANGELOG.md
@@ -13,12 +13,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed:
 
-- Adjusted several default rules
+- Adjusted several default rules:
   - Disabled `no-magic-numbers`
   - Corrected `import/no-unassigned-import` to allow for unnamed CSS imports
   - Fixed typo in module exports for react hooks rules
 - Moved JSX `ecmaFeatures` listing to React config and changed React version from fixed number to `detect`
-- Disabled the `sort-imports` rule as controls are not fine-grained enough, causing more problems than it solved.
+- Disabled the `sort-imports` rule as controls are not fine-grained enough, causing more problems than it solved
+- Add TypeScript support:
+  - Add TypeScript ESLint plugin and ruleset
+  - Update imports to resolve TypeScript files and use @typescript-eslint's parser
+  - Added `.ts` and `.tsx` to the `react/jsx-filename-extension` list
+  - Set `import/no-namespace` to off as \* imports are at times needed in typescript
+  - Add ESLint base rule overrides for TypeScript files
 - Explicitly disable deprecated rules
 - Switch from eslint-config-prettier to customized ruleset that matches Prettier's settings
 - Split the dependencies to be installed by users into peerDependencies for those absolutely necessary and optionalDependencies for those required by optional sub-packages